[chore] Added details section of wines and reconfigured the wines list component
